test_requirements.txt
black
bpython
codecov
ddt
django-debug-toolbar
factory_boy~=3.2
faker
freezegun==0.3.12
hypothesis==4.23.4
ipdb
isort
nplusone
pdbpp
pip-tools
pylint==2.6.0
pylint-django==2.3.0
pytest
pytest-cov
pytest-env==0.6.2
pytest-django
pytest-mock
pytest-lazy-fixture~=0.6.2
responses
semantic-version
wagtail-factories
